+++ hutool-core/src/main/java/cn/hutool/core/io/resource/FileResource.java
@@ -1,20 +1,23 @@
 package cn.hutool.core.io.resource;
 
-import java.io.File;
-import java.nio.file.Path;
-
 import cn.hutool.core.io.FileUtil;
-import cn.hutool.core.util.StrUtil;
 import cn.hutool.core.util.URLUtil;
 
+import java.io.File;
+import java.io.InputStream;
+import java.io.Serializable;
+import java.net.URL;
+import java.nio.file.Path;
+
 /**
- * 文件资源访问对象
+ * 文件资源访问对象，支持{@link Path} 和 {@link File} 访问
  * 
  * @author looly
- *
  */
-public class FileResource extends UrlResource {
+public class FileResource implements Resource, Serializable {
 	private static final long serialVersionUID = 1L;
+
+	private final File file;
 
 	// ----------------------------------------------------------------------- Constructor start
 	/**
@@ -43,7 +46,7 @@ public class FileResource extends UrlResource {
 	 * @param fileName 文件名，如果为null获取文件本身的文件名
 	 */
 	public FileResource(File file, String fileName) {
-		super(URLUtil.getURL(file), StrUtil.isBlank(fileName) ? file.getName() : fileName);
+		this.file = file;
 	}
 
 	/**
@@ -56,4 +59,36 @@ public class FileResource extends UrlResource {
 	}
 	// ----------------------------------------------------------------------- Constructor end
 
+	@Override
+	public String getName() {
+		return this.file.getName();
+	}
+
+	@Override
+	public URL getUrl(){
+		return URLUtil.getURL(this.file);
+	}
+
+	@Override
+	public InputStream getStream() throws NoResourceException {
+		return FileUtil.getInputStream(this.file);
+	}
+
+	/**
+	 * 获取文件
+	 *
+	 * @return 文件
+	 */
+	public File getFile() {
+		return this.file;
+	}
+
+	/**
+	 * 返回路径
+	 * @return 返回URL路径
+	 */
+	@Override
+	public String toString() {
+		return (null == this.file) ? "null" : this.file.toString();
+	}
 }