+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * 【core   】     增加compile包（pr#1243@Github）
 * 【core   】     增加ResourceClassLoader、CharSequenceResource、FileObjectResource
 * 【core   】     修改IoUtil.read(Reader)逻辑默认关闭Reader
+* 【core   】     ZipUtil增加Zip方法（pr#222@Gitee）
 
 ### Bug修复
 * 【cron   】     修复CronTimer可能死循环的问题（issue#1224@Github）